# Doctolib/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DocteurControlelr de l'api Rest",
        "description": "controlelr permettant de gérer les intéraction enter le client et la base de données",
        "version": "V1"
    },
    "paths": {
        "/docteurs": {
            "get": {
                "tags": [
                    "Tous les Ddocteurs"
                ],
                "summary": "Trouve l'ensemble des docteurs grace aux fonctions du repository",
                "description": "Retourne un tableau d'objet Docteur qui sera converti en tableau d'objets DocteurDTO ",
                "operationId": "App\\Controller\\DocteurRestController::searchAll",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocteurDTO"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocteurDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Erreur de requete"
                    },
                    "500": {
                        "description": "Nous rencontrons actuellement des problèmes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Créer un Docteur"
                ],
                "operationId": "App\\Controller\\DocteurRestController::create",
                "requestBody": {
                    "description": "DocteurDTO JSON Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocteurDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Input invalide"
                    },
                    "201": {
                        "description": "Création réussie"
                    }
                }
            }
        },
...
        "/patients/docteurs/{id}": {
            "get": {
                "tags": [
                    "Patients selon id Docteur"
                ],
                "summary": "Trouve l'ensemble des patients inscrits sur la bdd, ayant au moins un rendez vous avec le Docteur précisé par l'id dans l'URL",
                "description": "Retourne un tableau d'objets Patient qui sera converti en tableau d'objets PatientDTO ",
                "operationId": "App\\Controller\\PatientRestController::searchAllDocteurs",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du Docteur dont on cherche la liste des patients",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PatientDTO"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/PatientDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun docteur ne coreespond à votre requête"
                    },
                    "500": {
                        "description": "Nous rencontrons actuellement des problèmes"
                    }
                }
            }
        },
        "/patients/{id}": {
            "put": {
                "tags": [
                    "Modifier un Patient"
                ],
                "summary": "modification de objet Patient selon id",
                "description": "Ne peut être réalisée que par le Patient concerné",
                "operationId": "App\\Controller\\PatientRestController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du patient à modifier, qui est aussi l'id du patient connecté",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Mise à jour du Patient",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatientDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "Patient non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Supprimer un Patient"
                ],
                "summary": "Supprimer un patient",
                "description": "Uniquement accessible par le patient en question. L'ID se récupère automatiquement à la connexion du patient dans un $_SESSION",
                "operationId": "App\\Controller\\PatientRestController::remove",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id du patient à supprimer",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "Nous ne trouvons pas de patient avec cet id"
                    }
                }
            }
        },
        "/patients": {
            "post": {
                "tags": [
                    "Créer un Patient"
                ],
                "summary": "Création d'1 patient",
                "description": "Créationd u patient en inscription sur le site",
                "operationId": "App\\Controller\\PatientRestController::create",
                "requestBody": {
                    "description": "PatientDTO JSON Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatientDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Input invalide"
                    },
                    "201": {
                        "description": "Création réussie"
                    }
                }
            }
        },
...
        "/specialites": {
            "get": {
                "tags": [
                    "Toutes les specialites"
                ],
                "summary": "Trouve l'ensemble des specialite de la bdd",
                "description": "Retourne un tableau d'objet Docteur qui sera converti en tableau d'objets DocteurDTO. Accessible par tous les utilisateurs, connectés ou non. ",
                "operationId": "App\\Controller\\SpecialiteRestController::searchAll",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SpecialiteDTO"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/SpecialiteDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Erreur de requete"
                    },
                    "500": {
                        "description": "Nous rencontrons actuellement des problèmes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Créer une Specialite"
                ],
                "summary": "Création d'1 objet Specialite",
                "description": "Ne peut être réalisée que par des adminitrateurs",
                "operationId": "App\\Controller\\SpecialiteRestController::create",
                "requestBody": {
                    "description": "DocteurDTO JSON Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SpecialiteDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Champ mal renseigné"
                    },
                    "201": {
                        "description": "Création réussie"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "DocteurDTO": {
                "title": "DocteurDTO",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "number",
                        "format": "int64"
                    },
                    "username": {
                        "title": "username",
                        "description": "username, obligatoire",
                        "type": "string"
                    },
                    "password": {
                        "title": "password",
                        "description": "password, obligatoire",
                        "type": "string",
                        "format": "password"
                    },
                    "numeroOrdre": {
                        "title": "numeroOrdre",
                        "description": "numero d'ordre, unique et propre à chaque médecin, obligatoire",
                        "type": "string"
                    },
                    "nom": {
                        "title": "nom",
                        "description": "nom, obligatoire",
                        "type": "string"
                    },
                    "prenom": {
                        "title": "prenom",
                        "description": "prenom, obligatoire",
                        "type": "string"
                    },
                    "adresseTravail": {
                        "title": "adresseTravail",
                        "description": "adresse de travail du docteur, obligatoire",
                        "type": "string"
                    },
                    "codePostal": {
                        "title": "codePostal",
                        "description": "code postal de l'adresse de travail, obligatoire",
                        "type": "string"
                    },
                    "ville": {
                        "title": "ville",
                        "description": "ville de l'adresse de travail, obligatoire",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "description": "email du docteur, obligatoire",
                        "type": "string"
                    },
                    "telephone": {
                        "title": "téléphone",
                        "description": "téléphone professionnel du docteur",
                        "type": "string"
                    },
                    "lienSiteInternet": {
                        "title": "lien du site internet",
                        "description": "lien du site internet du docteur",
                        "type": "string"
                    },
                    "specialites": {
                        "title": "specialites",
                        "description": "tableaux des spécialités par docteur",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "priseRdvs": {
                        "title": "rdv",
                        "description": "tableaux des rdvs par docteurs",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DocteurDTO"
                }
            },
            "PatientDTO": {
                "title": "PatientDTO",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "number",
                        "format": "int64"
                    },
                    "username": {
                        "title": "username",
                        "description": "username, unique, obligatoire",
                        "type": "string"
                    },
                    "password": {
                        "title": "password",
                        "description": "password, hashé,obligatoire",
                        "type": "string"
                    },
                    "numeroCarteVitale": {
                        "title": "numero carte vitale",
                        "description": "numero de la carte vitale, unique également, obligatoire",
                        "type": "string",
                        "format": "password"
                    },
                    "nom": {
                        "title": "nom",
                        "description": "nom obligatoire",
                        "type": "string"
                    },
                    "prenom": {
                        "title": "prénom",
                        "description": "prénom obligatoire",
                        "type": "string"
                    },
                    "adresse": {
                        "title": "adresse",
                        "description": "adresse obligatoire",
                        "type": "string"
                    },
                    "ville": {
                        "title": "ville",
                        "description": "ville obligatoire",
                        "type": "string"
                    },
                    "codePostal": {
                        "title": "code postal",
                        "description": "code postal obligatoire",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "description": "email obligatoire",
                        "type": "string"
                    },
                    "telephone": {
                        "title": "telephone",
                        "description": "telephone pas obligatoire",
                        "type": "string"
                    },
                    "priseRdvs": {
                        "title": "rendez-vous",
                        "description": "tableaux des rendez vous du patient, les rendez vosu sont aussi des objets.",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PatientDTO"
                }
            },
...
            "SpecialiteDTO": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nom": {
                        "title": "nom",
                        "description": "libelle de la specialite, , obligatoire",
                        "type": "string"
                    },
                    "docteurs": {
                        "title": "docteurs",
                        "description": "tableaux des docteurs ayant cette spécialité",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        }
    }
}